+++ components/iris/include/iris_serde.h
@@ -10,7 +10,7 @@
 
 #include "driver/rmt.h"
 
-void serialize_data_from_ir_command(ir_command_t *, void **, int *);
+void serialize_data_from_ir_command(ir_command_t *, void **, size_t *);
 void deserialize_data_to_rmt_items(void *, unsigned int *, unsigned int *, unsigned int *, rmt_item32_t **);
